+++ src/commonMain/kotlin/at/posselt/pfrpg2e/data/kingdom/settlements/SettlementLevelUpType.kt
@@ -1,0 +1,24 @@
+package at.posselt.pfrpg2e.data.kingdom.settlements
+
+import at.posselt.pfrpg2e.data.ValueEnum
+import at.posselt.pfrpg2e.fromCamelCase
+import at.posselt.pfrpg2e.localization.Translatable
+import at.posselt.pfrpg2e.toCamelCase
+
+enum class SettlementLevelUpType : Translatable, ValueEnum {
+    TOWN,
+    CITY,
+    METROPOLIS,
+    METROPOLIS_THIRD_GRID,
+    METROPOLIS_FOURTH_GRID;
+
+    companion object {
+        fun fromString(value: String) = fromCamelCase<SettlementSizeType>(value)
+    }
+
+    override val value: String
+        get() = toCamelCase()
+
+    override val i18nKey: String
+        get() = "settlementLevelUpType.$value"
+}