+++ src/jsMain/kotlin/com/foundryvtt/core/documents/TokenConstraintOptions.kt
@@ -1,0 +1,9 @@
+package com.foundryvtt.core.documents
+
+import kotlinx.js.JsPlainObject
+
+@JsPlainObject
+external interface TokenConstraintOptions {
+    val ignoreCost: Boolean
+    val ignoreWalls: Boolean
+}