+++ src/jsMain/kotlin/com/foundryvtt/core/documents/TokenMovementSectionData.kt
@@ -1,0 +1,12 @@
+package com.foundryvtt.core.documents
+
+import kotlinx.js.JsPlainObject
+
+@JsPlainObject
+external interface TokenMovementSectionData {
+    val waypoints: Array<TokenMeasuredMovementWaypoint>
+    val distance: Int
+    val cost: Int
+    val spaces: Int
+    val diagonals: Int
+}